[[CHORE]] Move environment tests to envs.js
